Fix a casefolding bug. [64-bit]
In valid_utf8_p(), uc was improperly typed and never initialized. On
64-bit systems, this means that success is dependent on previous
stack contents.

If the upper 32 bits are not zero, the null terminator is not caught
and the function continues reading past the end of the string until:
1) Invalid UTF-8 is encountered
2) An invalid unicode codepoint is encountered.
3) segfault

1 and 2 are much more likely, but 3 is a danger.

+++ server/zstring.c
@@ -24,8 +24,8 @@ static String *zhash[STRING_HASH_TABLE_SIZE];
 
 int valid_utf8_p(const char* s)
 {
-    int len;
-    ssize_t uc;
+    ssize_t len;
+    int32_t uc;
 
     while ((len = utf8proc_iterate((const unsigned char *)s, -1, &uc))) {
         if (len <=0) return 0; /* Not valid UTF-8 encoding. */
